+++ e2e/wdio/platform/tests/thumbnail.e2e-spec.ts
@@ -63,28 +63,6 @@ describe('Thumbnail field', function() {
         }
     });
 
-    it('should highlight on hover', () => {
-        scrollIntoView(verticalGalleryImages);
-        const arrLength = getElementArrayLength(verticalGalleryImages);
-        for (let i = 0; arrLength > i; i++) {
-            scrollIntoView(verticalGalleryImages, i);
-            mouseHoverElement(verticalGalleryImages, i);
-            expect(getParentElementCSSProperty(verticalGalleryImages, 'border-bottom-color', i)).toContain('8,84,160');
-        }
-
-        for (let i = 0; arrLength > i; i++) {
-            scrollIntoView(horizontalGalleryImages, i);
-            mouseHoverElement(horizontalGalleryImages, i);
-            expect(getParentElementCSSProperty(horizontalGalleryImages, 'border-bottom-color', i)).toContain('8,92,175');
-        }
-
-        for (let i = 0; arrLength > i; i++) {
-            scrollIntoView(verticalGalleryVideo, i);
-            mouseHoverElement(verticalGalleryVideo, i);
-            expect(getParentElementCSSProperty(verticalGalleryVideo, 'border-bottom-color', i)).toContain('8,84,160');
-        }
-    });
-
     it('should be able to close gallery popup', () => {
         scrollIntoView(verticalGalleryImages, 4);
         waitForElDisplayed(verticalGalleryImages, 4);